feat: split Events into its own patch; rename calendar patch to "Hide LINE Calendar"
The calendar patch covered one feature (LINE Calendar) across five in-app
surfaces plus the unrelated chat-menu "Events" row. Separate them so each
patch maps to one feature:

- "Hide calendar button" -> renamed "Hide LINE Calendar"; scoped to the five
  Calendar surfaces only (header, toolbar, + attach tile, chat-menu row,
  message long-press menu).
- New "Hide Events button" patch (package line.hideevents) hides the chat-menu
  "Events" row (a shared d00.z item; the sole Lyz/s4;->l:Z read is replaced
  with const 0 at the build site so only that row drops).

Both are independent, default-on toggles. Verified by applying both against
LINE 26.11.0 and disassembling: 6 modified classes, all injections intact
(hg1.b.j returns false; no CALENDAR_BUTTON left in ed1.d0.a; d00.o isVisible
arg 0; ne1.y0$c.a returns null; s4.l read replaced by const 0 feeding the
Events row).
